+++ src/components/Addresses/NewAddress.tsx
@@ -4,7 +4,7 @@ import { useForm } from 'react-hook-form';
 import toAlpha2 from 'iso-3166-1-alpha-2';
 import { Context } from '../../utils/createContext';
 import { checkPostalCode } from '../../utils/validation';
-import { addAddress, updateCustomer } from '../../services/profileSetter';
+import { addAddress, updateCustomer } from '../../services/profileHandler/profileSetter';
 
 import '../../pages/Profile/Profile.scss';
