+++ src/pages/Login/Login.tsx
@@ -4,7 +4,7 @@ import { NavLink, Navigate, useNavigate } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
 import { ToastContainer, toast, Slide } from 'react-toastify';
 import { RoutesEnum } from '../../utils/enums';
-import { loginCustomer } from '../../services/customerAuther';
+import { loginCustomer } from '../../services/customerHandler/customerAuther';
 import { emailValidationRules, passwordValidationRules } from '../../utils/validation';
 import { Context } from '../../utils/createContext';
