+++ src/services/customerHandler/customerAuther.tsx
@@ -1,7 +1,7 @@
 import { createApiBuilderFromCtpClient, MyCustomerSignin, CustomerSignInResult } from '@commercetools/platform-sdk';
-import { loginUserCtpClient } from './withPasswordFlowClientBuilder';
-import { authClient } from './authAndHttpMiddlewareOptionsSetter';
-import { PROJECT_KEY, API_ADMIN_SCOPE } from './apiClientDetailsSetter';
+import { loginUserCtpClient } from '../flows/withPasswordFlowClientBuilder';
+import { authClient } from '../helpers/authAndHttpMiddlewareOptionsSetter';
+import { PROJECT_KEY, API_ADMIN_SCOPE } from '../helpers/apiClientDetailsSetter';
 
 export const loginCustomer = (request: MyCustomerSignin): Promise<CustomerSignInResult> => {
   const apiRoot = createApiBuilderFromCtpClient(loginUserCtpClient(request.email, request.password)).withProjectKey({
