+++ src/services/customerHandler/customerCreator.tsx
@@ -1,8 +1,8 @@
 import { CustomerSignInResult, createApiBuilderFromCtpClient } from '@commercetools/platform-sdk';
 import toAlpha2 from 'iso-3166-1-alpha-2';
-import { viewCustomersCtpClient, manageCustomersCtpClient } from './withClientCredentialsFlowClientBuilder';
-import { CustomerDraft, BaseAddress, EmailCheck } from '../utils/types';
-import { PROJECT_KEY } from './apiClientDetailsSetter';
+import { viewCustomersCtpClient, manageCustomersCtpClient } from '../flows/withClientCredentialsFlowClientBuilder';
+import { CustomerDraft, BaseAddress, EmailCheck } from '../../utils/types';
+import { PROJECT_KEY } from '../helpers/apiClientDetailsSetter';
 
 export const getCustomerDetails = (
   email: string,
