+++ src/services/flows/withAnonymousSessionFlowClientBuilder.tsx
@@ -2,8 +2,8 @@ import { ClientBuilder } from '@commercetools/sdk-client-v2';
 import {
   authMiddlewareOptionsForAnonymousSessionFlow,
   httpMiddlewareOptions,
-} from './authAndHttpMiddlewareOptionsSetter';
-import { API_ADMIN_SCOPE } from './apiClientDetailsSetter';
+} from '../helpers/authAndHttpMiddlewareOptionsSetter';
+import { API_ADMIN_SCOPE } from '../helpers/apiClientDetailsSetter';
 
 export const anonymousSessionFlowCtpClient = new ClientBuilder()
   .withAnonymousSessionFlow(authMiddlewareOptionsForAnonymousSessionFlow(API_ADMIN_SCOPE))
