+++ src/services/flows/withPasswordFlowClientBuilder.tsx
@@ -1,6 +1,9 @@
 import { ClientBuilder } from '@commercetools/sdk-client-v2';
-import { authMiddlewareOptionsForPasswordFlow, httpMiddlewareOptions } from './authAndHttpMiddlewareOptionsSetter';
-import { API_MANAGE_PROFILE_SCOPE } from './apiClientDetailsSetter';
+import {
+  authMiddlewareOptionsForPasswordFlow,
+  httpMiddlewareOptions,
+} from '../helpers/authAndHttpMiddlewareOptionsSetter';
+import { API_MANAGE_PROFILE_SCOPE } from '../helpers/apiClientDetailsSetter';
 
 export const loginUserCtpClient = (username: string, password: string) => {
   const ctpClient = new ClientBuilder()
