+++ src/services/helpers/authAndHttpMiddlewareOptionsSetter.tsx
@@ -7,7 +7,7 @@ import {
 } from '@commercetools/sdk-client-v2';
 import SdkAuth from '@commercetools/sdk-auth';
 import fetch from 'node-fetch';
-import { ApiClientDetails } from '../utils/types';
+import { ApiClientDetails } from '../../utils/types';
 import { PROJECT_KEY, REGION } from './apiClientDetailsSetter';
 
 export const authMiddlewareOptionsForClientCredentialsFlow = (details: ApiClientDetails): AuthMiddlewareOptions => {
